* commit '55eb6ae5959105ed4118a4c0f4fe886d4b687513':
  Программа кредитования

+++ src/main/java/kz/dilau/htcdatamanager/domain/CreditProgramm.java
@@ -21,8 +21,8 @@ public class CreditProgramm extends BaseEntity<Long> {
     private String nameRu;
     private String nameKz;
     private String nameEn;
-    private BigDecimal minDownPayment;
-    private BigDecimal maxDownPayment;
+    private Long minDownPayment;
+    private Long maxDownPayment;
     private BigDecimal minCreditPeriod;
     private BigDecimal maxCreditPeriod;
     private BigDecimal percent;
